refactor: move "channels.ts" logic into "channels/outbound.ts"
