+++ integrations/teams/src/index.ts
@@ -1,6 +1,6 @@
 import { sentry as sentryHelpers } from '@botpress/sdk-addons'
 import actions from './actions'
-import { channels } from './channels'
+import { channels } from './channels/outbound'
 import { register, unregister } from './setup'
 import { handler } from './webhooks/handler'
 import * as bp from '.botpress'
